Begin creating unit test functions and modify polarTransform mapping function
Setup basic structure for unit tests. The basic method will be to input an image and compare that to a presaved image that one must verify is correct.

Alter convertToPolarImage to use SciPy map_coordinate function instead of OpenCV remap. First, this makes me feel better because it is more Pythonic. Additionally, there is no need to convert to 32-bit floats to perform computation.