Clean up how images are generated for tests and fix bug with initial/final angle
The original idea was to delete the main.py file when I was finished with tests but I don't think that is a great idea. Instead, I renamed it to testsGeneration and moved into the tests folder. Each image that is generated is wrote into a function that can be called in the file to regenerate it easily. In addition, a developer can add debugging plots to show the image and ensure it looks right.

In addition, a bug was fixed with the initial and final angle and radii so that now they are submaps from the entire image.