Switch from OpenCV remapping in cartesian conversion
This was previously done in the polar conversion with switching from using OpenCV's remap function to using SciPy's map_coordinates function. This is more Pythonic where SciPy is specifically built for Python while OpenCV is adapted to Python. In addition, it is not required to convert to 32-bit floats to be able to run the function.