Fix issue with corrupt edge
The problem was that the mode of 'constant' when converting to polar made an error when converting back to cartesian because it would average out the pixel value with the outside value in the image.

The answer is to pad the image with edges. In this case, a 3px border is added all around to compensate for this. In the future, this needs to be done only if the border value is set to constant. Otherwise if it is set to something else, this will affect those results.