Remove origin parameter from polar and cartesian conversion
I was conflicted on whether to keep the parameter. A few of the issues that the parameter brought:
First, I would have to appropriately change the center and points to take into account the origin. This is a lot of additional work.

Second, and the biggest problem by far, is that it only makes sense to have the polar image be oriented such that the origin is in the lower left hand corner.

If you look at numpy and corresponding scientific packages, they do not have anything regarding origin. My realization is that the origin is just used for displaying the image at the end. It can be up to the user if they want to use the upper or lower origin configuration for storing images. They just flip the image if upper and it will then be fine.

If this issue gets enough attention then I can revisit it but for now it is not worth the trouble