Remove initial/final source radius/angle from cartesian conversion
The original purpose for these parameters were to select particular regions of the polar image to render on the screen. It quickly became difficult to manage when adding image orientations (middle-middle, top-middle, etc) and the capability of automatically determining the image size when none is given.

I don't necessarily see a specific need for this yet but it is nice to have the option to give specifications when converting back to cartesian domain.

Since the initial/final source radius/angle will essentially change the image size and initial/final angle/radius, I am thinking maybe a function named clipPolarImage where you can specify these parameters and it will clip the image and return you the image along with the new metadata to use.

This can then be passed along to the cartesian conversion function and it will work as expected. This idea works well but I am not going to bother with it until someone comes along with a need for this functionality.