Remove initial/final source radius/angle from cartesian conversion
The original purpose for these parameters were to select particular regions of the polar image to render on the screen. It quickly became difficult to manage when adding image orientations (middle-middle, top-middle, etc) and the capability of automatically determining the image size when none is given.

I don't necessarily see a specific need for this yet but it is nice to have the option to give specifications when converting back to cartesian domain.

Since the initial/final source radius/angle will essentially change the image size and initial/final angle/radius, I am thinking maybe a function named clipPolarImage where you can specify these parameters and it will clip the image and return you the image along with the new metadata to use.

This can then be passed along to the cartesian conversion function and it will work as expected. This idea works well but I am not going to bother with it until someone comes along with a need for this functionality.

Remove origin parameter from polar and cartesian conversion

I was conflicted on whether to keep the parameter. A few of the issues that the parameter brought:
First, I would have to appropriately change the center and points to take into account the origin. This is a lot of additional work.

Second, and the biggest problem by far, is that it only makes sense to have the polar image be oriented such that the origin is in the lower left hand corner.

If you look at numpy and corresponding scientific packages, they do not have anything regarding origin. My realization is that the origin is just used for displaying the image at the end. It can be up to the user if they want to use the upper or lower origin configuration for storing images. They just flip the image if upper and it will then be fine.

If this issue gets enough attention then I can revisit it but for now it is not worth the trouble